Controller to service (#8)
* moved controller code to service and fixed a bug in userService

* changed adminchart service/portfolio to be like the rest

* changed whitespace to tabs

* refactored code to get the code to run

* removed findById in UserRepo
