+++ src/main/java/com/forge/revature/controllers/ExceptionController.java
@@ -1,7 +1,6 @@
 package com.forge.revature.controllers;
 
 import com.forge.revature.exception.NotFoundException;
-
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,12 +11,12 @@ import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExcep
 
 @ControllerAdvice
 public class ExceptionController extends ResponseEntityExceptionHandler{
-  
-  //if something not found, gives 404 response
-  @ExceptionHandler({ NotFoundException.class })
-  protected ResponseEntity<Object> handleNotFound(Exception ex, WebRequest request) {
-    return handleExceptionInternal(ex, ex.getMessage(), 
-      new HttpHeaders(), HttpStatus.NOT_FOUND, request);
-  }
+	
+	//if something not found, gives 404 response
+	@ExceptionHandler({ NotFoundException.class })
+	protected ResponseEntity<Object> handleNotFound(Exception ex, WebRequest request) {
+		return handleExceptionInternal(ex, ex.getMessage(), 
+		new HttpHeaders(), HttpStatus.NOT_FOUND, request);
+	}
 }
