+++ src/main/java/com/forge/revature/controllers/FullPortfolioIgnoreMixin.java
@@ -2,13 +2,12 @@ package com.forge.revature.controllers;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.forge.revature.models.User;
-
 import java.util.HashMap;
 
 public abstract class FullPortfolioIgnoreMixin {
-    @JsonIgnore
-    User user;
-
-    @JsonIgnore
-    HashMap<String, String> flags;
+	@JsonIgnore
+	User user;
+	
+	@JsonIgnore
+	HashMap<String, String> flags;
 }
