+++ src/test/java/com/forge/revature/demo/PortfolioTest.java
@@ -54,6 +54,7 @@ import com.forge.revature.repo.ProjectRepo;
 import com.forge.revature.repo.UserRepo;
 import com.forge.revature.repo.WorkExperienceRepo;
 import com.forge.revature.repo.WorkHistoryRepo;
+import com.forge.revature.services.PortfolioService;
 import com.forge.revature.repo.MatrixRepo;
 import com.forge.revature.repo.SkillRepo;
 
@@ -106,9 +107,10 @@ public class PortfolioTest {
 	@BeforeEach
 	public void setup() {
 		mvc = MockMvcBuilders.standaloneSetup(
-				new PortfolioController(repo, aboutMeRepo, certificationRepo, educationRepo, equivalencyRepo,
-						gitHubRepo, honorRepo, projectRepo, workExperienceRepo, workHistoryRepo, matrixRepo, skillRepo))
+				new PortfolioController(new PortfolioService(repo, aboutMeRepo, certificationRepo, educationRepo, equivalencyRepo, 
+						gitHubRepo, honorRepo, projectRepo, workExperienceRepo, workHistoryRepo, matrixRepo, skillRepo)))
 				.build();
+		
 	}
 
 	@Test
